refactor(services/mysql): split service into standalone crate
