+++ core/Cargo.toml
@@ -139,7 +139,7 @@ services-mini-moka = ["opendal-core/services-mini-moka"]
 services-moka = ["dep:opendal-service-moka"]
 services-mongodb = ["opendal-core/services-mongodb"]
 services-monoiofs = ["opendal-core/services-monoiofs"]
-services-mysql = ["opendal-core/services-mysql"]
+services-mysql = ["dep:opendal-service-mysql"]
 services-obs = ["opendal-core/services-obs"]
 services-onedrive = ["opendal-core/services-onedrive"]
 services-opfs = ["opendal-core/services-opfs"]
@@ -203,6 +203,7 @@ opendal-service-azfile = { path = "services/azfile", version = "0.55.0", optiona
 opendal-service-ghac = { path = "services/ghac", version = "0.55.0", optional = true, default-features = false }
 opendal-service-hdfs-native = { path = "services/hdfs-native", version = "0.55.0", optional = true, default-features = false }
 opendal-service-moka = { path = "services/moka", version = "0.55.0", optional = true, default-features = false }
+opendal-service-mysql = { path = "services/mysql", version = "0.55.0", optional = true, default-features = false }
 opendal-service-s3 = { path = "services/s3", version = "0.55.0", optional = true, default-features = false }
 
 [dev-dependencies]
