+++ core/services/mysql/src/lib.rs
@@ -15,28 +15,21 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use std::sync::Arc;
+/// Default scheme for mysql service.
+pub const MYSQL_SCHEME: &str = "mysql";
 
-use super::core::*;
-use crate::raw::oio;
-use crate::raw::*;
-use crate::*;
+use opendal_core::DEFAULT_OPERATOR_REGISTRY;
 
-pub struct MysqlDeleter {
-    core: Arc<MysqlCore>,
-    root: String,
-}
+mod backend;
+mod config;
+mod core;
+mod deleter;
+mod writer;
 
-impl MysqlDeleter {
-    pub fn new(core: Arc<MysqlCore>, root: String) -> Self {
-        Self { core, root }
-    }
-}
+pub use backend::MysqlBuilder as Mysql;
+pub use config::MysqlConfig;
 
-impl oio::OneShotDelete for MysqlDeleter {
-    async fn delete_once(&self, path: String, _: OpDelete) -> Result<()> {
-        let p = build_abs_path(&self.root, &path);
-        self.core.delete(&p).await?;
-        Ok(())
-    }
+#[ctor::ctor]
+fn register_mysql_service() {
+    DEFAULT_OPERATOR_REGISTRY.register::<Mysql>(MYSQL_SCHEME);
 }
